refactor(playback): extract weighted random sampler and artwork warmer services
- Implement WeightedPlaybackSampler to encapsulate preference scoring, runtime penalties, and weighted random track selection.
- Implement PlaybackArtworkWarmer to isolate neighboring queue track artwork preloading/warm-up logic.
- Simplify PlaybackCoordinator by delegating to WeightedPlaybackSampler and PlaybackArtworkWarmer.
