+++ kmgccc_player/Services/Playback/PlaybackCoordinator.swift
@@ -30,8 +30,7 @@ final class PlaybackCoordinator {
     private var lastTelemetryIsPlaying: Bool?
     private var sidecarHydrationTask: Task<Void, Never>?
     private var sidecarHydratingTrackID: UUID?
-    private var trackArtworkWarmupTask: Task<Void, Never>?
-    private var trackArtworkWarmupSignature: String?
+    private let artworkWarmer: PlaybackArtworkWarmer
     private let lyricSnippetSeekLeadInSeconds: Double = 0.8
 
     private struct LyricsRefetchContext: Equatable {
@@ -54,13 +53,15 @@ final class PlaybackCoordinator {
         appleMusicAdapter: AppleMusicPlaybackAdapter,
         systemNowPlayingProvider: SystemNowPlayingProvider,
         settings: AppSettings? = nil,
-        meterProvider: AudioLevelMeterProtocol? = nil
+        meterProvider: AudioLevelMeterProtocol? = nil,
+        artworkWarmer: PlaybackArtworkWarmer = PlaybackArtworkWarmer()
     ) {
         self.playerVM = playerVM
         self.appleMusicAdapter = appleMusicAdapter
         self.systemNowPlayingProvider = systemNowPlayingProvider
         self.settings = settings ?? AppSettings.shared
         self.meterProvider = meterProvider
+        self.artworkWarmer = artworkWarmer
         self.activeSource = PlaybackSource(
             rawValue: UserDefaults.standard.string(forKey: Keys.activeSource) ?? ""
         ) ?? .local
@@ -429,9 +430,9 @@ final class PlaybackCoordinator {
     }
 
     func playRandomTracks(_ tracks: [Track], libraryQueueSource: PlayerViewModel.LibraryQueueSource? = nil) {
-        let queue = Self.availableUniqueTracks(from: tracks)
+        let queue = WeightedPlaybackSampler.playableUniqueTracks(from: tracks)
         guard !queue.isEmpty else { return }
-        let startTrack = Self.smartRandomPick(from: queue) ?? queue[0]
+        let startTrack = WeightedPlaybackSampler.pick(from: queue) ?? queue[0]
         let startIndex = queue.firstIndex(where: { $0.id == startTrack.id }) ?? 0
         playTracks(
             queue,
@@ -446,7 +447,7 @@ final class PlaybackCoordinator {
         inQueueFrom tracks: [Track],
         libraryQueueSource: PlayerViewModel.LibraryQueueSource? = nil
     ) {
-        let queue = Self.availableUniqueTracks(from: tracks)
+        let queue = WeightedPlaybackSampler.playableUniqueTracks(from: tracks)
         guard !queue.isEmpty else { return }
         let startIndex = queue.firstIndex(where: { $0.id == track.id }) ?? 0
         playTracks(
@@ -462,7 +463,7 @@ final class PlaybackCoordinator {
         seekTo seconds: Double,
         libraryQueueSource: PlayerViewModel.LibraryQueueSource? = nil
     ) {
-        let queue = Self.availableUniqueTracks(from: tracks)
+        let queue = WeightedPlaybackSampler.playableUniqueTracks(from: tracks)
         guard !queue.isEmpty else { return }
         let startIndex = queue.firstIndex(where: { $0.id == track.id }) ?? 0
         playTracks(
@@ -540,7 +541,11 @@ final class PlaybackCoordinator {
         }
 
         notifyTelemetryIfNeeded(source: activeSource, isPlaying: isPlaying)
-        scheduleTrackArtworkWarmupIfNeeded(for: newPresentation)
+        artworkWarmer.update(
+            activeSource: activeSource,
+            presentation: newPresentation,
+            queue: playerVM.currentQueueTracks
+        )
 
         let previousPresentation = presentation
         guard !newPresentation.isEffectivelyEqual(to: previousPresentation) else { return }
@@ -555,67 +560,14 @@ final class PlaybackCoordinator {
     @available(*, deprecated, message: "Use smartRandomPick for single picks or playRandomTracks for ShuffleSession-backed playback.")
     static func smartRandomQueue(from tracks: [Track], startingWith startTrack: Track? = nil) -> [Track] {
         if let startTrack,
-           let matched = availableUniqueTracks(from: tracks).first(where: { $0.id == startTrack.id }) {
+           let matched = WeightedPlaybackSampler.playableUniqueTracks(from: tracks).first(where: { $0.id == startTrack.id }) {
             return [matched]
         }
-        return smartRandomPick(from: tracks).map { [$0] } ?? []
+        return WeightedPlaybackSampler.pick(from: tracks).map { [$0] } ?? []
     }
 
     static func smartRandomPick(from tracks: [Track]) -> Track? {
-        let uniqueTracks = availableUniqueTracks(from: tracks)
-        guard !uniqueTracks.isEmpty else { return nil }
-
-        let trackByID = Dictionary(uniqueKeysWithValues: uniqueTracks.map { ($0.id, $0) })
-        let weights = Dictionary(uniqueKeysWithValues: uniqueTracks.map { track in
-            let stats = PreferenceStatsService.shared.getStats(for: track.id)
-            let score = PreferenceScorerV2.calculateScore(
-                stats: stats,
-                duration: track.duration,
-                manualLikeState: stats.manualLikeState
-            )
-            return (track.id, score.baseWeight)
-        })
-
-        return weightedSample(
-            from: uniqueTracks,
-            weights: weights,
-            recentHistory: [],
-            trackByID: trackByID
-        )
-    }
-
-    private static func availableUniqueTracks(from tracks: [Track]) -> [Track] {
-        var seenIDs = Set<UUID>()
-        return tracks.filter { track in
-            guard !seenIDs.contains(track.id) else { return false }
-            seenIDs.insert(track.id)
-            return track.availability != .missing
-        }
-    }
-
-    private static func weightedSample(
-        from tracks: [Track],
-        weights: [UUID: Double],
-        recentHistory: [UUID],
-        trackByID: [UUID: Track]
-    ) -> Track? {
-        let adjustedWeights = Dictionary(uniqueKeysWithValues: tracks.map { track in
-            let baseWeight = weights[track.id] ?? 1
-            let runtimeWeight = PreferenceScorerV2.applyRuntimePenalties(
-                baseWeight: baseWeight,
-                track: track,
-                recentHistory: recentHistory,
-                tracks: trackByID
-            )
-            return (track.id, runtimeWeight)
-        })
-
-        guard let selectedID = WeightedRandomSampler.sample(
-            from: tracks.map(\.id),
-            weights: adjustedWeights
-        ) else { return nil }
-
-        return trackByID[selectedID]
+        WeightedPlaybackSampler.pick(from: tracks)
     }
 
     private func externalProvider(for source: PlaybackSource) -> (any ExternalPlaybackProvider)? {
@@ -770,53 +722,6 @@ final class PlaybackCoordinator {
             track.artworkFileName ?? "no-file",
             ArtworkDataFingerprint.sampledString(for: artworkData),
         ].joined(separator: ":")
-    }
-
-    private func scheduleTrackArtworkWarmupIfNeeded(for presentation: NowPlayingPresentation) {
-        guard activeSource == .local, let currentTrack = presentation.localTrack else {
-            trackArtworkWarmupTask?.cancel()
-            trackArtworkWarmupTask = nil
-            trackArtworkWarmupSignature = nil
-            return
-        }
-
-        let queue = playerVM.currentQueueTracks
-        let currentIndex = queue.firstIndex(where: { $0.id == currentTrack.id })
-        var targets: [Track] = [currentTrack]
-        if let currentIndex {
-            for offset in 1...2 {
-                let index = currentIndex + offset
-                if queue.indices.contains(index) {
-                    targets.append(queue[index])
-                }
-            }
-            let previousIndex = currentIndex - 1
-            if queue.indices.contains(previousIndex) {
-                targets.append(queue[previousIndex])
-            }
-        }
-
-        var seen = Set<UUID>()
-        let sources = targets.compactMap { track -> TrackArtworkSource? in
-            guard seen.insert(track.id).inserted else { return nil }
-            let fallbackData = track.id == currentTrack.id ? presentation.artworkData : track.artworkData
-            return track.trackArtworkSource(fallbackData: fallbackData)
-        }
-        guard !sources.isEmpty else {
-            trackArtworkWarmupTask?.cancel()
-            trackArtworkWarmupTask = nil
-            trackArtworkWarmupSignature = nil
-            return
-        }
-
-        let signature = sources.map(\.sourceKey).joined(separator: "||")
-        guard signature != trackArtworkWarmupSignature else { return }
-        trackArtworkWarmupSignature = signature
-        trackArtworkWarmupTask?.cancel()
-        trackArtworkWarmupTask = TrackArtworkCache.shared.preloadPlaybackArtwork(
-            for: sources,
-            reason: "playback-current-window"
-        )
     }
 
     private func textSignature(_ text: String?) -> String {
